refactor: change imports, delete unnecessary files
