--- app/src/main/kotlin/view/neo4jInput.kt
@@ -1,9 +1,0 @@
-package view
-
-data class Neo4jInput(
-    var uri: String = "",
-    var login: String = "",
-    var password: String = "",
-    var isUpdated: Boolean = false,
-    var isUndirected: Boolean = false
-)
